Quad mesh interpolation
Fixes to Field2D interpolation from quad meshes. 1) modified the
interpolation algorithm to use statistical "dart-throwing". This helps
us get data in cut cells that previously would have zero value on the
external node. 2) added a sanity check to XtoL on quad meshes. Still
getting some incorrect results with some quads so this makes sure the
computed l,m coordinate actually agrees with the input position. If not,
brute force search is launched which is slow so this is only a temporary
work around.